fix: candidate and roulette struct moved
